more re-sync with netcdf4
netcdf.h has changed a lot in the 4.x series.  Since it's now possible for
pnetcdf to be inclucded as part of netcdf, it's important both projects have
the same values for the flags.   this commit partialy re-syncs with netcdf4,
though we still have to re-sync the limits, which have been extended in recent
netcdf4 releases 


git-svn-id: https://svn.mcs.anl.gov/repos/parallel-netcdf/trunk@917 c368ceb8-fd28-0410-a457-e46ee1d7cd4a