The original design of nonblocking I/O is to concatenate the fileviews of individual nonblocking requests and serve them with a single MPI-IO call.  However, because MPI-IO requires the file displacements of the flattened fileview be monotonically nondecreasing, we divide the nonblocking requests into groups such that each group abides by the fileview requirement. Then, each group is carried out by a separate MPI-IO call. Performance can be poor if there are too many groups.
This revision fixes this problem by flattening all fileviews into offset-length pairs and merging all the pairs, so that the nonblocking requests can be carried out by a single MPI-IO call. Performance is expected to be improved as the number of MPI-IO calls is reduced to one.

However, be warned about the additional memory requirement. The additional memory needed for flatterning the fileviews might be more than the I/O data itself. For example, each offset-length pair represents only a 4-byte integer and the C struct for the offset-length pair takes 3*sizeof(MPI_Offset)=24 bytes.



git-svn-id: https://svn.mcs.anl.gov/repos/parallel-netcdf/trunk@1121 c368ceb8-fd28-0410-a457-e46ee1d7cd4a