Remove some tests for MPI functions and constants
As PnetCDF requires MPI-IO which was first introduced in MPI 2,
we can simply check MPI_VERSION defined in all mpi.h to skip
tests for MPI functions and constants introduced in MPI 2.
Check for deprecated MPI functions and constants are only needed
when the supplied MPI libraries support MPI 3 and later, i.e.
those functions and constants may be still defined in the libraries.