Replace use of PATH_MAX with MPI_MAX_INFO_VAL
Burst buffer directory name is passed from an MPI hint and the argument
"value" of MPI_Info_get() is limited to the size of MPI_MAX_INFO_VAL.
The default of PATH_MAX is 4096. This value can be bigger than
MPI_MAX_INFO_VAL and thus we cannot use PATH_MAX.

For MPICH, the the default value of MPI_MAX_INFO_VAL is 1024, which
cannot be changed even at configure time.

For OpenMPI, the default value of MPI_MAX_INFO_VAL is 256, but can
be changed at configure time through command-line option
--with-max-info-val=VALUE.