add implementation to ignore file layout order of variables
NetCDF file format specification contains a comment below about the
file space locations of non-record variables in the data section:

   non_recs = [vardata ...] // The data for all non-record variables,
                            // stored contiguously for each variable, in
                            // the same order the variables occur in the
                            // header.

The similar comment does not appear for record variables.

This requirement is actually not necessary, as each variable has a
'begin' entity in the header pointing to the starting file location
storing its data. Removing this requirement has an advantage when
file header needs to expand. It allows to move only the first few
variables large enough for the required additional header space to
locations with high file offsets, effectively reducing the expensive
data movement. Note we cannot change the variable order already in the
file header, as applications may already hold variable IDs that point to
the variable array indices in the header.

However, this relaxation may result in the earlier PnetCDF and NetCDF
versions complaining the file format error with NC_ENOTNC.