Performance tuning: for nonblocking requests of more than one record
The original design to keep the pending list of nonblocking requests
sorted based on the variable's begin offsets is to avoid an expensive
sorting at ncmpi_wait_all. However, the design only assumes all pending
requests of record variables are for the same record, and thus the
sorting uses the beginnings of the record variables (first records).
This design was reasonable, as usually the application would flush all
pending requests to one record before posting requests to the next
record. Therefore, when nonblocking requests consist of multiple
records, inserting the new requests of next record into the pending list
becomes expensive. This commit improves the design by using the
beginnings of variable's individual records.