test case from E3SM
This test case once uncovered a memory leak in ROMIO.  Would like to
make it a ROMIO test, but because it uses pnetcdf routines it should
live here.  See https://github.com/pmodels/mpich/issues/2800 for more
context.

modified to create the small dataset as well as read from it